(revert) Legacy: Revert GuiTest and remove ApplicationTest, FxTest and CustomAppTest.
